requirements.txt
absl-py==0.9.0
asn1crypto==1.3.0
astor==0.8.1
atari-py==0.2.6  # TODO(hartikainen): This is weirdly required but not installed by gym
attrs==19.3.0
boto3==1.11.5
botocore==1.14.5
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
dm-control==0.0.300771433
dm-env==1.2
dm-tree==0.1.1
docutils==0.15.2
fasteners==0.15
filelock==3.0.12
funcsigs==1.0.2
future==0.18.2
gast>=0.3.2
gitdb2==2.0.6
GitPython==3.1.0
glfw==1.9.1
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gtimer==1.0.0b5
gym==0.15.4
h5py==2.10.0
idna==2.7
imageio==2.6.1
importlib-metadata==1.4.0
jmespath==0.9.4
ipdb==0.12.3
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
lz4==3.0.2
Markdown==3.1.1
matplotlib==3.1.2
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
monotonic==1.5
more-itertools==8.1.0
moviepy==1.0.3
mujoco-py==2.0.2.9
networkx==2.4
numpy==1.17.5
oauthlib==3.1.0
opencv-python==4.1.2.30
opencv-python-headless==4.1.2.30
opt-einsum==3.1.0
packaging==20.0
pandas==0.25.3
pathlib2==2.3.5
Pillow==7.0.0
pip==19.3.1
pluggy==0.13.1
protobuf==3.11.2
psutil==5.6.7
py==1.8.1
py-spy==0.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==2.7.3
pycosat==0.6.3
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.3
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
ray[tune]==0.8.4
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
robosuite==0.1.0
rsa==4.0
s3transfer==0.3.1
scikit-learn==0.22.1
scikit-image==0.16.2
scikit-video==1.1.11
scipy==1.4.1
setproctitle==1.1.10
setuptools==45.1.0
six==1.13.0
smmap2==2.0.5
tabulate==0.8.6
tensorboard==2.2.0
tensorflow==2.2.0rc3
tensorflow-addons==0.9.1
tensorflow-estimator==2.2.0
tensorflow-probability==0.10.0rc0
termcolor==1.1.0
tqdm==4.41.1
urllib3==1.24.3
wcwidth==0.1.8
Werkzeug==0.16.0
wheel==0.33.6
wrapt==1.11.2
zipp==1.0.0
